feat: implement task monitoring API integration with real-time updates
- Add comprehensive task API proxy routes (/api/tasks, /api/tasks/[id], /api/tasks/[id]/action, /api/tasks/events)
- Create TaskEventsClient for Server-Sent Events streaming of real-time task updates
- Update task API client to use real V3 backend endpoints with proper parameter mapping
- Implement task filtering, pagination, and action controls via API
- Add SSE proxy for real-time task event streaming with error handling
- Update TODO status to reflect partial task monitoring completion
- Add comprehensive task lifecycle management (create, update, actions, events)

BREAKING: Task monitoring now connects to real V3 backend APIs instead of mock data, enabling live agent task tracking and control

+++ iterations/v3/indexers/src/database.rs
@@ -39,12 +39,12 @@ impl DatabasePool {
 
     /// Get number of active connections
     pub fn num_idle(&self) -> usize {
-        self.pool.num_idle() as usize
+        self.pool.num_idle()
     }
 
     /// Get pool size
     pub fn size(&self) -> usize {
-        self.pool.size() as usize
+        self.pool.size()
     }
 }
